Refactor getTodos pipeline to use Html () not String
Html () is a stronger type and the String was encoding things
weirdly as it is not meant to be rendered as HTML. It escaped all
the HTML characters like '<' and '>', meaning the browser would
render the HTML literally instead of making a list, etc. as desired.

Also figured out the right combination pof lucid packages.