Add a 'mouseentered' test route
It is interesting that this type-safe routing system would not even
render the part of the code that relies on posting to a route
that did not exist.

<div hx-post="/mouse_entered" hx-trigger="mouseenter">[Here Mouse, Mouse!]</div>

With raw WAI, I'd just get a runtime 404. I think I still did get a 404,
but it did not render the "[Here Mouse, Mouse!]" string at all on the
page, which it was doing before hte type safe routes.

This composable, type-safe routes concept is nice!