Refactor Main into Todo.hs Routes.hs and Database.hs
There's a lot of duplication for now. Some functionality is not ported,
but the prototype of patterns is in place.

The homepage shows what it needs, but does not load '/todos'.
The /todos route shows the list, but cannot be added to.

Adding a new list item to homepage form turns the url into:
  localhost:3000/?msg=Hello+World
for entry string of "Hello World".

This is not the old behavior. Will need to finish porting and
investigate.