Extend routes to accept request header
Is this the correct approach?

I need the request headers to be in scope of the functions that need
them. In this case, that is `postForm`, whch relies on the list of
params in the request body.

While this `req` passing seems a bit like a hack, it is important to
realize the scope of `wai-inv-routes` package is limited to
unopinionated route creation (the parameters inside the route string).

This other stuff I am doing is related to the application logic  `app`.

The fact that wai-inv-route makes its own simplified version of the
`Request` header and provides an interop function to move between
Network.Wai.Request -> Network.Wai.Invertible.Request is evidence of
this.

-- |Convert a 'Wai.Request' to a request.
waiRequest :: Wai.Request -> Request
waiRequest q = Request
  { requestHost = maybe [] splitHost $ Wai.requestHeaderHost q
  , requestSecure = Wai.isSecure q
  , requestMethod = toMethod $ Wai.requestMethod q
  , requestPath = Wai.pathInfo q
  , requestQuery = simpleQueryParams $ map (second $ fromMaybe mempty) $ Wai.queryString q
  , requestContentType = fromMaybe mempty $ lookup hContentType headers
  } where headers = Wai.requestHeaders
.

As you can see below, the Wai.Request type is large and includes the
equestBody` field that Invertible strips (because it doesn't need those
fields).

-- Wai.Request typedef
data Request = Request {
     requestMethod        :: H.Method
  ,  httpVersion          :: H.HttpVersion
  ,  rawPathInfo          :: B.ByteString
  ,  rawQueryString       :: B.ByteString
  ,  requestHeaders       :: H.RequestHeaders
  ,  isSecure             :: Bool
  ,  remoteHost           :: SockAddr
  ,  pathInfo             :: [Text]
  ,  queryString          :: H.Query
  ,  requestBody          :: IO B.ByteString
  ,  vault                 :: Vault
  ,  requestBodyLength     :: RequestBodyLength
  ,  requestHeaderHost     :: Maybe B.ByteString
  ,  requestHeaderRange   :: Maybe B.ByteString
  ,  requestHeaderReferer   :: Maybe B.ByteString
  ,  requestHeaderUserAgent :: Maybe B.ByteString
  }
  deriving (Typeable)
.

The Wai.getRequestBodyChunk function literally operates on the missing
field. Thus, this is in fact the correct approach.

getRequestBodyChunk :: Request -> IO B.ByteString
getRequestBodyChunk = requestBody